refactor: layout footer header change dir
